* moves UsersComponent into dedicated component folder
